requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Aadhaar OCR — Tesseract + OpenCV (fully local)
pytesseract==0.3.13
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# Document extraction — Gemini Vision (REST, SSL disabled)
google-generativeai==0.8.3
httpx==0.27.0

# Config
python-dotenv==1.0.1